Export attach() also as a default
Previously, `attach()` was exported as _whole_ of the package.  However,
`index.d.ts` says that it is exported as _default_ of package.  _whole_
and _default_ is not the same thing in ES2015 and TypeScript, they are
always confusing:

```javascript
// Whole import
import * as attach from 'neovim-client';

// Default import
import attach from 'neovim-client'
```

In general, people prefers default export because it's much more
generic.  So I exported `attach()` also as a 'default' of this package.