debug: spawn child process with --inspect-brk
npm uses shims (cmd.exe, sh, node) in node_modules/.bin/
to run a nodejs script from a npm module.
On Ubuntu, there is 1 shim for nodejs.
On Windows, there is 1 shim each for cmd.exe and sh.
npm shims do not allow the user to pass arguments to nodejs
and debugging a module requires passing arguments to nodejs
so the user can debug a module only by running the nodejs directly
and passing to it the filepath of the nodejs script, referred by the shim.
Determining this filepath is not fast enough for npm and yarn
so the client should handle this.

The client spawns nodejs with '--inspect-brk' to run a child client,
preserving all commandline arguments, under the following conditions:

1. $NVIM_NODE_HOST_DEBUG is set.
2. nodejs supports '--inspect-brk'. Check nodejs semver version.
3. '--inspect-brk' is not passed to nodejs.

Fix #46

NodeJS Changelog links:
https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V6.md#6.12.0
https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V7.md#7.6.0
https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V8.md#8.0.0