fix(nvim.js): $NVIM_LISTEN_ADDRESS is deprecated
Use $NVIM with Nvim 0.8+.

Since https://github.com/neovim/neovim/pull/11009 (Nvim 0.8+) Nvim does
_not_ pass $NVIM_LISTEN_ADDRESS to child processes.  The correct way to
detect a "host" or "parent" Nvim is to check for $NVIM env var: this
means the current process is a child of Nvim.

The old $NVIM_LISTEN_ADDRESS env var had conflicting purposes as both
a parameter ("the current process should _listen on_ this address") and
a descriptor ("the current process is a _child of_ this address").